% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktshares.R
\name{mktshares}
\alias{mktshares}
\title{Compute market share in an airport or city}
\usage{
mktshares(dataframe, carrier, location)
}
\arguments{
\item{dataframe}{Dataset to be based market share off of, should have a Passengers column}

\item{carrier}{Column name containing carrier name of each ticket, unquoted}

\item{location}{Column name containing the origination, unquoted}
}
\value{
a dataframe with the carrier and orig columns and a market_share
column that contains the market share of that carrier in that location,
as a proportion.
}
\description{
Compute market share in an airport or city, with flexibility to specify how the
carrier is defined
}
\examples{
\dontshow{if (require("tibble")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
require("tibble")
data = tibble:: tribble(
     ~origin, ~airline, ~Passengers,
* "CLT", "American," 10,
"RDU","American", 5,
"RDU", "United", 7
)

market_shares(data, airline, carrier)
\dontshow{\}) # examplesIf}
}
